requirements.txt
Flask
pytesseract
Pillow
opencv-python
tiktoken
gunicorn
openai==0.28.1
celery
redis